requirements.txt
django-crispy-forms==1.14.0
django-recaptcha==3.0.0
django-two-factor-auth==1.15.2
django-two-factor-auth[phonenumbers]==1.15.2
django-auto-logout==0.5.0
pytz==2022.4
django-environ==0.9.0
django-axes==5.39.0